Restore githubScraping.ts. Fix small issues
