+++ packages/graphql_server/src/test/githubApi.test.ts
@@ -1,7 +1,7 @@
 import { getRepositoryTopics } from '../api/githubApi'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-async function testGetRepositoryTopics(repoFounder: string, repoName: string) {
+export async function testGetRepositoryTopics(repoFounder: string, repoName: string) {
   //get the real comments from method in hackerNewsScraper. E.g. searchHackerNewsStories('SuperAGI')
 
-  console.log(await getRepositoryTopics(repoFounder, repoName, 'token'))
+  console.log(await getRepositoryTopics(repoFounder, repoName, process.env.GITHUB_API_TOKEN))
 }
