+++ packages/graphql_server/src/test/hackerNewsScraping.test.ts
@@ -1,7 +1,7 @@
 import { searchHackerNewsStories } from '../scraping/hackerNewsScraping'
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-async function testHackerNewsScraping(repoName: string) {
-  //returns the comments
+export async function testHackerNewsScraping(repoName: string) {
+  //logs the comments
   console.log(await searchHackerNewsStories(repoName))
 }
